proy_api_recarg/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.7,<4.0
djangorestframework-simplejwt
django-simple-history
django-filter
drf-yasg
django-guardian
django-rest-auth
Pillow
psycopg2
